# api/docs/swagger.yaml
definitions:
  models.Car:
    properties:
      crated_at:
        type: string
      daily_limit:
        type: integer
      id:
        type: string
      investor_id:
        type: string
      investor_percentage:
        type: number
      km:
        type: integer
      model:
        type: string
      over_limit:
        type: integer
      price:
        type: number
      state_number:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.Client:
    properties:
      address:
        type: string
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
  models.CreateCar:
    properties:
      daily_limit:
        type: integer
      investor_id:
        type: string
      investor_percentage:
        type: number
      km:
        type: integer
      model:
        type: string
      over_limit:
        type: integer
      price:
        type: number
      state_number:
        type: string
    type: object
  models.CreateClient:
    properties:
      address:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
...
  models.Empty:
    type: object
  models.GetListCarResponse:
    properties:
      cars:
        items:
          $ref: '#/definitions/models.Car'
        type: array
      count:
        type: integer
    type: object
  models.GetListClientResponse:
    properties:
      clients:
        items:
          $ref: '#/definitions/models.Client'
        type: array
      count:
        type: integer
    type: object
...
  models.UpdateCarSwag:
    properties:
      daily_limit:
        type: integer
      investor_id:
        type: string
      investor_percentage:
        type: number
      km:
        type: integer
      model:
        type: string
      over_limit:
        type: integer
      price:
        type: number
      state_number:
        type: string
      status:
        type: string
    type: object
  models.UpdateClientSwag:
    properties:
      address:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /car:
    get:
      consumes:
      - application/json
      description: Get List Car
      operationId: get_list_car
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetCarListBody
          schema:
            $ref: '#/definitions/models.GetListCarResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Car
      tags:
      - Car
    post:
      consumes:
      - application/json
      description: Create Car
      operationId: create_car
      parameters:
      - description: CreateCarRequestBody
        in: body
        name: Car
        required: true
        schema:
          $ref: '#/definitions/models.CreateCar'
      produces:
      - application/json
      responses:
        "201":
          description: GetCarBody
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Car
      tags:
      - Car
  /car/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Car
      operationId: delete_car
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteCarBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Car
      tags:
      - Car
    get:
      consumes:
      - application/json
      description: Get By ID Car
      operationId: get_by_id_car
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetCarBody
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get By ID Car
      tags:
      - Car
    put:
      consumes:
      - application/json
      description: Update Car
      operationId: update_car
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateCarRequestBody
        in: body
        name: Car
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCarSwag'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateCarBody
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Car
      tags:
      - Car
  /client:
    get:
      consumes:
      - application/json
      description: Get List Client
      operationId: get_list_client
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetClientListBody
          schema:
            $ref: '#/definitions/models.GetListClientResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Client
      tags:
      - Client
    post:
      consumes:
      - application/json
      description: Create Client
      operationId: create_client
      parameters:
      - description: CreateClientRequestBody
        in: body
        name: Client
        required: true
        schema:
          $ref: '#/definitions/models.CreateClient'
      produces:
      - application/json
      responses:
        "201":
          description: GetClientBody
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Client
      tags:
      - Client
  /client/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Client
      operationId: delete_client
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteClientBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Client
      tags:
      - Client
    get:
      consumes:
      - application/json
      description: Get By ID Client
      operationId: get_by_id_client
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetClientBody
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get By ID Client
      tags:
      - Client
    put:
      consumes:
      - application/json
      description: Update Client
      operationId: update_client
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateClientRequestBody
        in: body
        name: Client
        required: true
        schema:
          $ref: '#/definitions/models.UpdateClientSwag'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateClientBody
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Client
      tags:
      - Client
...
swagger: "2.0"